+++ apps/server/src/env.ts
@@ -1,4 +1,4 @@
-import { flyRegions } from "@openstatus/db/src/schema/shared";
+import { flyRegions } from "@openstatus/db/src/schema/constants";
 import { createEnv } from "@t3-oss/env-core";
 import { z } from "zod";
