+++ packages/db/src/schema/monitors/validation.ts
@@ -5,7 +5,7 @@ import * as assertions from "@openstatus/assertions";
 
 import { monitorJobTypes, monitorMethods, monitorStatus } from "./constants";
 import { monitor, monitorsToPages } from "./monitor";
-import { monitorPeriodicitySchema, monitorRegionSchema } from "../shared";
+import { monitorPeriodicitySchema, monitorRegionSchema } from "../constants";
 
 export const monitorMethodsSchema = z.enum(monitorMethods);
 export const monitorStatusSchema = z.enum(monitorStatus);
