+++ packages/plans/src/index.ts
@@ -1,7 +1,3 @@
 export * from "./config";
 export * from "./utils";
 export * from "./pricing-table";
-export * from "./version";
-
-export { workspacePlans as plans } from "@openstatus/db/src/schema";
-export type { WorkspacePlan } from "@openstatus/db/src/schema";
